requirements.txt
# Web framework and server
fastapi==0.115.12
uvicorn==0.34.2

# Google Generative AI
google-genai>=1.14.0,<1.15.0

# Async HTTP client
aiohttp>=3.8.4

# Additional utilities
pytz==2025.2
python-multipart>=0.0.6
websockets>=11.0.3

# For handling CORS
starlette>=0.27.0

# Logging and environment
python-dotenv>=1.0.0

pipecat-ai[daily,google,assembly,silero,openai,azure,elevenlabs]
pipecat-ai[noisereduce]

# Google Cloud Speech
google-cloud-speech

# Misc
loguru

# OpenTelemetry for Langfuse tracing
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp-proto-http
langfuse

twilio
pipecat-ai-flows

# Redis for message queue and caching
redis>=5.0.0

# System monitoring
psutil>=5.9.0
